+++ packages/@runlightyear/airtable/src/index.ts
@@ -6,45 +6,96 @@ export type { AirtableProps } from "./Airtable";
 export { AirtableOAuth } from "./AirtableOAuth";
 export type { AirtableOAuthProps } from "./AirtableOAuth";
 
-/* Types */
-export type { AirtableScope } from "./types/AirtableScope";
-export type { Timezone } from "./types/Timezone";
+/* Bases */
+export type {
+  GetBaseSchemaProps,
+  GetBaseSchemaResponse,
+} from "./bases/getBaseSchema";
+export type {
+  ListBasesProps,
+  ListBasesResponse,
+  ListBasesResponseData,
+} from "./bases/listBases";
 
-/* Who am i */
+/* Listeners */
+export type {
+  OnChangeProps,
+  OnChangeRunFunc,
+  OnChangeRunFuncProps,
+} from "./listeners/onChange";
+export type {
+  OnNewOrUpdatedRecordsProps,
+  OnNewOrUpdatedRecordsRunFunc,
+  OnNewOrUpdatedRecordsRunFuncProps,
+} from "./listeners/onNewOrUpdatedRecords";
+export type {
+  OnNewRecordsProps,
+  OnNewRecordsRunFunc,
+  OnNewRecordsRunFuncProps,
+} from "./listeners/onNewRecords";
+
+/* Meta */
 export type { WhoamiResponse } from "./meta/whoami";
 
-/* Create Records */
+/* Records */
 export type {
   CreateRecordsProps,
   SingleRecordResponseData,
   MultipleRecordsResponseData,
   CreateRecordsResponse,
 } from "./records/createRecords";
-
-/* Delete Record */
 export type {
   DeleteRecordProps,
   DeleteRecordResponse,
   DeleteRecordResponseData,
 } from "./records/deleteRecord";
-
-/* Get Record */
 export type {
   GetRecordProps,
   GetRecordResponse,
   GetRecordResponseData,
 } from "./records/getRecord";
-
-/* List Records */
 export type {
   ListRecordsProps,
   ListRecordsResponse,
   ListRecordsResponseData,
 } from "./records/listRecords";
-
-/* Update Record */
 export type {
   UpdateRecordProps,
   UpdateRecordResponse,
   UpdateRecordResponseData,
 } from "./records/updateRecord";
+
+/* Types */
+export type { AirtableScope } from "./types/AirtableScope";
+export type { Timezone } from "./types/Timezone";
+
+/* Webhooks */
+export type {
+  CreateWebhookProps,
+  CreateWebhookResponse,
+  CreateWebhookResponseData,
+} from "./webhooks/createWebhook";
+export type {
+  AirtableWebhookSubscribeProps,
+  AirtableWebhookSubscribePropsFunc,
+  DefineAirtableWebhookProps,
+} from "./webhooks/defineAirtableWebhook";
+export type {
+  DeleteWebhookProps,
+  DeleteWebhookResponse,
+} from "./webhooks/deleteWebhook";
+export type {
+  ListWebhookPayloadsProps,
+  ListWebhookPayloadsResponse,
+  ListWebhookPayloadsResponseData,
+} from "./webhooks/listWebhookPayloads";
+export type {
+  ListWebhooksProps,
+  ListWebhooksResponse,
+  ListWebhooksResponseData,
+} from "./webhooks/listWebhooks";
+export type {
+  RefreshWebhookProps,
+  RefreshWebhookResponse,
+  RefreshWebhookResponseData,
+} from "./webhooks/refreshWebhook";