+++ packages/@runlightyear/airtable/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @runlightyear/airtable
 
+## 0.7.0
+
+### Minor Changes
+
+- adb48382d: Support for refreshing expiring webhook subscriptions
+- adb48382d: listWebhooks method bugfix
+
+### Patch Changes
+
+- Updated dependencies [4393656c3]
+  - @runlightyear/lightyear@0.36.0
+
 ## 0.6.0
 
 ### Minor Changes