+++ packages/@runlightyear/github/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/github
 
+## 0.21.1
+
+### Patch Changes
+
+- Updated dependencies [4393656c3]
+  - @runlightyear/lightyear@0.36.0
+
 ## 0.21.0
 
 ### Minor Changes