+++ packages/@runlightyear/gsheets/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/gsheets
 
+## 0.5.2
+
+### Patch Changes
+
+- Updated dependencies [4393656c3]
+  - @runlightyear/lightyear@0.36.0
+
 ## 0.5.1
 
 ### Patch Changes