+++ packages/@runlightyear/lightyear/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @runlightyear/lightyear
 
+## 0.36.0
+
+### Minor Changes
+
+- 4393656c3: Support for refreshing expiring webhook subscriptions
+
 ## 0.35.0
 
 ### Minor Changes