+++ packages/@runlightyear/notion/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/notion
 
+## 0.4.2
+
+### Patch Changes
+
+- Updated dependencies [4393656c3]
+  - @runlightyear/lightyear@0.36.0
+
 ## 0.4.1
 
 ### Patch Changes