+++ packages/@runlightyear/zoom/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/zoom
 
+## 0.5.6
+
+### Patch Changes
+
+- Updated dependencies [4393656c3]
+  - @runlightyear/lightyear@0.36.0
+
 ## 0.5.5
 
 ### Patch Changes